Renaming files to ship with es5 / es6 versions of the project
